Refactor lib_expense
Add data layer
